### 2019-02-22 #1271 @whedon
[PRE REVIEW]: JuliaCall: an R package for seamless integration between R and Julia
**Submitting author:** @Non-Contradiction (<a href="http://orcid.org/0000-0001-7022-0947">Changcheng Li</a>)
**Repository:** <a href="https://github.com/Non-Contradiction/JuliaCall" target ="_blank">https://github.com/Non-Contradiction/JuliaCall</a>
**Version:** v0.16.4
