### 2019-02-25 #1271 @leouieda
@arfon sure, I'd be happy to do it :+1:

I'll look through the reviewer list to see if anyone has R and Julia experience. @Non-Contradiction let me know if you have any suggestions for reviewers. There is [a list of potential reviewers here](https://bit.ly/joss-reviewers).
