### 2019-02-27 #1271 @leouieda
:wave: Hi @ChrisRackauckas @Aguang @matbesancon @r-barnes @nhejazi @dmbates would any of you be interested in reviewing this submission for JOSS?
