### 2019-02-27 #1271 @leouieda
@matbesancon @aguang  @Non-Contradiction the review will take place in #1284. Please feel free to ping me at any time if you have questions.

@matbesancon @aguang there are instructions in the issue and a link to our review guidelines. When you open issues and/or pull requests in the software repo, please link back to the review issue so we can keep track of progress by including `openjournals/joss-reviews#1284` in the text.
