### 2019-02-28 Non-Contradiction/JuliaCall#85 @matbesancon
[JOSS] License file is not MIT
A full license file is required in the repo, see github templates for MIT license or here: https://opensource.org/licenses/MIT
