### 2019-02-28 Non-Contradiction/JuliaCall#85 @Non-Contradiction
Thank you for the feedback!

The license file currently follows the CRAN template for MIT license instead of a full license.
For example, see <https://github.com/r-lib/R6/blob/master/LICENSE>, <https://github.com/r-lib/pkgdown/blob/master/LICENSE>, and <https://github.com/r-lib/evaluate/blob/master/LICENSE> for some license of popular R packages under MIT.

The CRAN requires that we write a line like https://github.com/r-lib/R6/blob/67f20c53f39ef34c1f049e9f1248588d725ab685/DESCRIPTION#L20 in the Description file and use a license file that contains years and copyright holders only.

Can I just mention the package is MIT licensed in something like README or paper itself?
I think I can deal with this problem by having different license files on CRAN and Github, but this solution is not ideal...
