### 2019-02-28 ebb3a46
Updates in README for the issue of libstdc++.so.6 [skip ci].
