+++ README.Rmd
@@ -18,12 +18,12 @@ output: github_document
 - [JuliaCall for Seamless Integration of R and Julia](#juliacall-for-seamless-integration-of-r-and-julia)
   - [Installation](#installation)
   - [Basic Usage](#basic-usage)
-  - [Trouble Shooting and Way to Get Help](#trouble-shooting-and-way-to-get-help)
+  - [Troubleshooting and Way to Get Help](#troubleshooting-and-way-to-get-help)
   - [JuliaCall for R Package Developers](#juliacall-for-r-package-developers)
   - [Suggestion and Issue Reporting](#suggestion-and-issue-reporting)
-  - [Interfaces between R and Julia](#interfaces-between-r-and-julia)
+  - [Other Interfaces Between R and Julia](#other-interfaces-between-r-and-julia)
 
-Package `JuliaCall` is an R interface to `Julia`,
+Package `JuliaCall` is an R interface to `Julia`, 
 which is a high-level, high-performance dynamic programming language
 for numerical computing, see <https://julialang.org/> for more information.
 Below is an image for [Mandelbrot set](https://en.wikipedia.org/wiki/Mandelbrot_set).
@@ -42,14 +42,13 @@ knitr::opts_chunk$set(
 
 ## Installation
 
-To use package `JuliaCall`, you first have to install [`Julia`](https://julialang.org/) on your computer, you can download a generic binary from 
+To use package `JuliaCall`, you first have to install [`Julia`](https://julialang.org/) on your computer. You can download a generic binary from 
 <https://julialang.org/downloads/> and add it to your path, and then you can install `JuliaCall` just like any other R packages by
 ```{r eval=FALSE}
 install.packages("JuliaCall")
 ```
 
-Note that currently `Julia v0.6.x`, `Julia v0.7.0` and `Julia v1.0` are all supported by `JuliaCall`.
-But note that some `Julia` packages may not work on `Julia v1.0` and you may need to use `Julia v0.7.0` for transition purposes.
+Currently `Julia v0.6.x`, `Julia v1.0.x`, and `Julia v1.1` are all supported by `JuliaCall`.
 
 You can get the development version of `JuliaCall` by
 ```{r eval=FALSE}
@@ -65,11 +64,11 @@ library(JuliaCall)
 
 julia <- julia_setup()
 
-## If you want to use a julia at a specific location, you could do the following:
-## julia_setup(JULIA_HOME = "the folder that contains julia binary"), 
+## If you want to use Julia at a specific location, you could do the following:
+## julia_setup(JULIA_HOME = "the folder that contains Julia binary"), 
 ## or you can set JULIA_HOME in command line environment or use `options(...)`
 
-## Different ways for using Julia to calculate sqrt(2)
+## Different ways of using Julia to calculate sqrt(2)
 
 # julia$command("a = sqrt(2);"); julia$eval("a")
 julia_command("a = sqrt(2);"); julia_eval("a")
@@ -93,31 +92,35 @@ julia_installed_package("Optim")
 julia_library("Optim")
 ```
 
-## Trouble Shooting and Way to Get Help
+## Troubleshooting and Way to Get Help
 
 ### Julia is not found
 
 Make sure the `Julia` installation is correct.
-`JuliaCall` is able to find `Julia` on PATH,
+`JuliaCall` can find `Julia` on PATH,
 and there are three ways for `JuliaCall` to find `Julia` not on PATH.
 
 - Use `julia_setup(JULIA_HOME = "the folder that contains julia binary")`
 - Use `options(JULIA_HOME = "the folder that contains julia binary")`
 - Set `JULIA_HOME` in command line environment.
 
+### libstdc++.so.6: version `GLIBCXX_3.4.xx' not found
+
+Such problems are usually on Linux machines.
[…]
