### 2019-02-28 Non-Contradiction/JuliaCall#86 @matbesancon
julia_setup() deprecated Pkg use
On `julia_setup()` run:
```
running command ''/home/user/programs/julia/bin/julia' '/home/user/R/x86_64-pc-linux-gnu-library/3.4/JuliaCall/julia/install_dependency.jl' '/usr/lib/R' 2>&1' had status 1
```

I think this is due to `Pkg.dir` is deprecated in https://github.com/Non-Contradiction/JuliaCall/blob/master/inst/julia/install_dependency.jl
