### 2019-02-28 Non-Contradiction/JuliaCall#87 @matbesancon
How is the test suite run?
I might have missed a step:
```R
library(testthat)
test_check("JuliaCall")
Error: No tests found for JuliaCall
test_package("JuliaCall")
Error: No tests found for JuliaCall
```
