### 2019-02-28 Non-Contradiction/JuliaCall#87 @Non-Contradiction
I think to test the package, it is better to deal with the source package instead of the installed package. Since the installed package may not have tests, vignettes, and etc.
To run the test suit for the source package
1. download the GitHub repo,
2. open `JuliaCall.Rproj` if you have RStudio or open R from the root of the downloaded directory,
3. run `devtools::test()` for the test suite.
Note that you may need to run `devtools::test()` multiple times to finish the test suit because of an issue in `devtools`, and the first time of the test can be quite long.
