### 2019-02-28 Non-Contradiction/JuliaCall#87 @matbesancon
ok I'll try these steps. This should be in the README also
