### 2019-02-28 Non-Contradiction/JuliaCall#86 @Non-Contradiction
Thank you for the feedback.

The `Pkg.dir` may not be the reason for the status 1. But the status 1 definitely means something is not right...

The use of the deprecated `Pkg.dir` is intended.
The purpose is to check whether rebuilding JuliaCall's dependency is needed or not,
and the new function which replaces the deprecated `Pkg.dir` is quite time-consuming.
Hence I keep the `Pkg.dir`.

But after thinking the problem a bit more, since rebuilding is not common, which should only happen when R is updated to a new version, I may remove this checking in `julia_setup()`, and have a separate function like `rebuild()` which contains the checking and rebuilding, and users can use it only when needed. Then the new function `rebuild()` can use the more time-consuming checking function which is not deprecated.
