### 2019-02-28 Non-Contradiction/JuliaCall#87 @Non-Contradiction
There is something about these steps in the README section "suggestion and issue reporting", but I think I need to change the wording a little bit. Thanks for the suggestion!
