### 2019-02-28 d44b776
Updates in README about the checking and testing of the package. [skip ci]
