### 2019-02-28 Non-Contradiction/JuliaCall#87 @Non-Contradiction
There is now a clear subsection in README about the testing of the package!
