### 2019-03-01 Non-Contradiction/JuliaCall#86 @matbesancon
it's a good idea yes, reading the readme, it's not clear what the function of setup() is and how often this should be performed
