### 2019-03-01 e85b01e
Add LICENSE.md for the full MIT license. Related to #85.
