### 2019-03-01 Non-Contradiction/JuliaCall#85 @Non-Contradiction
I follow the example of https://github.com/ropensci/emld, which is also an R package under MIT license and has been accepted to JOSS recently, and create a [LICENSE.md](https://github.com/Non-Contradiction/JuliaCall/blob/master/LICENSE.md) in the repository.

Hope this fulfills the requirement of a full license file in the repo.
Thank you very much!
