### 2019-03-01 Non-Contradiction/JuliaCall#86 @Non-Contradiction
The `julia_setup()` is necessary for each new R session to use the package. It will do all the necessary setup, like setting up the display system if necessary and loading `RCall.jl` in the embedded Julia, which is for type conversion between R and Julia, etc. And it will be invoked automatically before carrying out other `JuliaCall` functions. I will make this clear in the README.

I will also start a new development branch for the idea of `rebuild()` function, and it will be contained in the next release of `JuliaCall`.
