### 2019-03-01 769a848
More clear about the functionality of `julia_setup()` in README. [skip ci]
