+++ README.Rmd
@@ -60,7 +60,9 @@ devtools::install_github("Non-Contradiction/JuliaCall")
 ```{r}
 library(JuliaCall)
 
-## Do initial setup
+## Do initial setup for automatic type conversion, Julia display systems, etc. 
+## It is necessary for every new R session to use the package.
+## If not carried out manually, it will be invoked automatically before other julia_xxx functions.
 
 julia <- julia_setup()
 
