### 2019-03-01 08993e3
Split RCall checking, related to #86
Split RCall checking from install_dependency into a new `rebuild` function.
