### 2019-03-01 0c48230
Add the new rebuild function, related to #86
