### 2019-03-01 23a194c
Add rebuild argument in function julia_setup to trigger RCall's rebuild, related to #86.
