### 2019-03-01 Non-Contradiction/JuliaCall#86 @Non-Contradiction
The deprecated Pkg usage has been removed in #88, and I add a new argument `rebuild` in the `julia_setup`  function, whose default value is `FALSE` and setting it to `TRUE` will trigger the rebuild of `RCall.jl` if necessary.
I also update the documentation of `julia_setup` to make it more clear.
The pull request changes the functionality of the package, and it will be included in the next release.
