### 2019-03-03 Non-Contradiction/JuliaCall#89 @matbesancon
[JOSS] Contribution guidelines
The README indicates how to get help and file issue, but not guidelines for people who want to contribute (how can people help, are there some coding guidelines, should each PR correspond to at least one issue, etc)
