### 2019-03-03 580cc25
Add code_of_conduct and contributing guide, related to #89. [skip ci]
