### 2019-03-03 1d9f388
Fix links of the issue template and the pull request template. [skip ci]
