+++ README.Rmd
@@ -146,8 +146,8 @@ you may report it using the link: <https://github.com/Non-Contradiction/JuliaCal
 `JuliaCall` is under active development now.
 Any suggestion or issue reporting is welcome!
 You may report it using the link: <https://github.com/Non-Contradiction/JuliaCall/issues/new>, or email me at <lch34677@gmail.com> or <cxl508@psu.edu>.
-You are welcome to use the [issue template](.github/issue-template.md)
-and the [pull request template](.github/pull-request-template.md).
+You are welcome to use the [issue template](.github/issue_template.md)
+and the [pull request template](.github/pull_request_template.md).
 The [contributing guide](.github/CONTRIBUTING.md) gives some guidance to make contributions.
 
 ### Checking `JuliaCall` Package
