### 2019-03-03 Non-Contradiction/JuliaCall#89 @Non-Contradiction
Thanks! I added a code of conduct and contributing guide following and adapted from ropensci's advice at <https://ropensci.github.io/dev_guide/collaboration.html#friendlyfiles>
