### 2019-03-04 #1284 @matbesancon
@aguang I'm almost finished reviewing the software part and going to start the paper itself soon
