### 2019-03-04 #1284 @leouieda
Thanks for the update @matbesancon :+1:

If you have feedback or require changes to the software/docs, please open issues or PRs in the software repo and link them here by mentioning `openjournals/joss-reviews#1284`.
