### 2019-03-04 #1284 @matbesancon
Closed:
https://github.com/Non-Contradiction/JuliaCall/issues/85
https://github.com/Non-Contradiction/JuliaCall/issues/89

Still open:
https://github.com/Non-Contradiction/JuliaCall/issues/87
https://github.com/Non-Contradiction/JuliaCall/issues/86
