### 2019-03-04 Non-Contradiction/JuliaCall#85 @leouieda
Linking to openjournals/joss-reviews#1284
