### 2019-03-05 a4f5615
Try to use valgrind in testing.
