### 2019-03-05 Non-Contradiction/JuliaCall#91 @aguang
`julia_setup()` cannot find Julia
This may be related to the many other issues that appear to be open with `julia_setup()`. When I run it with no arguments as described in **Basic Usage** I get that Julia is not found. It is fine when I supply my own path, but this suggests that `julia_setup()` has issues actually finding my Julia installation.

```r
> julia <- julia_setup()
sh: NA: command not found
Error in julia_setup() : Julia is not found.
In addition: Warning message:
In system2("bash", "-l -c 'which julia'", stdout = TRUE) :
  running command ''bash' -l -c 'which julia'' had status 1

> julia <- julia_setup(JULIA_HOME= "/Applications/Julia-1.1.app/Contents/Resources/julia/bin/")
Julia version 1.1.0 at location /Applications/Julia-1.1.app/Contents/Resources/julia/bin will be used.
Loading setup script for JuliaCall...
Finish loading setup script for JuliaCall.
```

<details> <summary><strong>Session Info</strong></summary>

> sessionInfo()
R version 3.5.1 (2018-07-02)
Platform: x86_64-apple-darwin15.6.0 (64-bit)
Running under: macOS  10.14.3

Matrix products: default
BLAS: /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/3.5/Resources/lib/libRlapack.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:

other attached packages:
[1] JuliaCall_0.16.4

[1] compiler_3.5.1 tools_3.5.1    yaml_2.2.0     Rcpp_1.0.0     knitr_1.21     xfun_0.5
