### 2019-03-05 a6275a6
Do not use r_check_args [skip ci]
As Valgrind is not available on mac os.
