+++ .travis.yml
@@ -1,17 +1,13 @@
 language: r
 cache: packages
-r_check_args: '--use-valgrind'
 os:
   - linux
   - osx
 env:
   - JULIA="06"
-    VALGRIND_OPTS='--leak-check=full --track-origins=yes'
   # - JULIA="07"
   - JULIA="10"
-    VALGRIND_OPTS='--leak-check=full --track-origins=yes'
   - JULIA="11"
-    VALGRIND_OPTS='--leak-check=full --track-origins=yes'
 
 addons:
   apt:
@@ -47,5 +43,3 @@ before_install:
   - export PATH="$HOME/julia/bin:$PATH"
 
   - if [ "$TRAVIS_OS_NAME" == "linux" ]; then export R_LD_LIBRARY_PATH="$(R RHOME)/lib:$HOME/julia/lib/julia"; fi
-
-  # - if [ $JULIA == "10" ]; then julia -e 'using Pkg; pkg"add RCall#master"'; fi
