### 2019-03-05 Non-Contradiction/JuliaCall#91 @Non-Contradiction
Thanks for the feedback!

It seems that there is no Julia installation on PATH, and the Julia installation can't be invoked by just `julia` in the terminal, which is why there is the warning message `running command ''bash' -l -c 'which julia'' had status 1`.
This is the current design of `julia_setup`, which either looks for Julia at the location provided by the user or looks for Julia in the PATH, which is a reasonable solution plausible on all os systems (but not ideal). Also, it is documented in README and the R package documentation.

An ideal solution is to try to find Julia harder in a system specific way if the current mechanism fails. For example, `JuliaCall` may look at `/Applications/` folder on mac for Julia, and it can invoke that Julia if there is only one Julia at paths like `/Applications/Julia-x.x.app` (since it is possible to have multiple versions of Julia at the same time).

I will try to implement the ideal solution step by step when I have time. Thank you for the suggestion!
