### 2019-03-06 Non-Contradiction/JuliaCall#91 @aguang
That sounds fine to me. I would also suggest either

  1) documenting that `julia_setup()` will look within PATH earlier in the **Basic Usage** section, so that if it can't find it it means that Julia is not within your PATH (for example, I use the [alias solution](https://en.wikibooks.org/wiki/Introducing_Julia/Getting_started) to run from terminal with OSX), or
  2) referencing the Troubleshooting section in **Basic Usage** for any errors with `julia_setup()`.

I understand that this seems to be a OSX specific issue, since all other OSes will allow you to run from command line immediately.
