### 2019-03-06 Non-Contradiction/JuliaCall#92 @aguang
Documenting difference between `julia_call`, `julia_eval`, `julia_command`
Would you be able to document somewhere the difference between `julia_call`, `julia_eval`, and `julia_command`? In particular the difference between these functions in returning a value to R and  returning a value to Julia.

It took me a while to understand that `julia_command` does not return a value to R, but will display return values from Julia (if `show_value=TRUE`) while `julia_eval` will not display return values from Julia but will return a value to R (if `need_return %in% c("R","Julia")`) while `julia_call` can do either/both.
