### 2019-03-06 Non-Contradiction/JuliaCall#93 @aguang
'julia_call` appears to have trouble with typing
When I try to run the `convert` function using `julia_call`, I get a `MethodError` from Julia that I need to provide a `Type` as the first argument. This does not occur when I run using `julia_command` or `julia_eval`. I am able to fix the error by providing `julia_eval("Float32")` as the argument instead.

```{r}
> julia_command("convert(Float32,1/3)")
0.33333334f0

> julia_call("convert","Float32","1/3")
 Show Traceback

 Rerun with Debug
 Error: Error happens in Julia.
MethodError: First argument to `convert` must be a Type, got Float32
Stacktrace:
 [1] docall(::Ptr{Nothing}) at /Users/aguang/CORE/scratch/JuliaCall/inst/julia/setup.jl:168

> julia_eval("convert(Float32,1/3)")
[1] 0.3333333

> julia_call("convert", julia_eval("Float32"),1/3)
[1] 0.3333333
```

Similarly, I have been testing JuliaCall out with the BioSequences package following along [their examples](https://biojulia.net/BioSequences.jl/stable/sequences/bioseq.html). When I attempt to use `julia_call` to run certain functions I will get `MethodError: no method matching xyz` where the arguments all appear to have the type String:

```{r}
> julia_install_package("BioSequences")
> julia_library("BioSequences")
> julia_command('dna"TACGTANNATC"')
11nt DNA Sequence:
TACGTANNATC
> julia_command('DNASequence(dna"ACGT", dna"NNNN", dna"TGCA")')
12nt DNA Sequence:
ACGTNNNNTGCA

> julia_call("DNASequence",'dna"ACGT"','dna"NNNN"','dna"TGCA"')
 Show Traceback

 Rerun with Debug
 Error: Error happens in Julia.
MethodError: no method matching BioSequence{DNAAlphabet{4}}(::String, ::String, ::String)
Closest candidates are:
  BioSequence{DNAAlphabet{4}}(::Union{AbstractString, AbstractArray{T,1} where T}) where A<:Alphabet at /Users/aguang/.julia/packages/BioSequences/7i86L/src/bioseq/constructors.jl:29
  BioSequence{DNAAlphabet{4}}(::Union{AbstractString, AbstractArray{T,1} where T}, !Matched::Integer) where A<:Alphabet at /Users/aguang/.julia/packages/BioSequences/7i86L/src/bioseq/constructors.jl:29
  BioSequence{DNAAlphabet{4}}(::Union{AbstractString, AbstractArray{T,1} where T}, !Matched::Integer, !Matched::Integer) where A<:Alphabet at /Users/aguang/.julia/packages/BioSequences/7i86L/src/bioseq/constructors.jl:29
Stacktrace:
 [1] docall(::Ptr{Nothing}) at /Users/aguang/CORE/scratch/JuliaCall/inst/julia/setup.jl:168
> julia_call("repeat", 'dna"TA"',10)
 Show Traceback

 Rerun with Debug
 Error: Error happens in Julia.
MethodError: no method matching repeat(::String, ::Float64)
Closest candidates are:
  repeat(::Union{SubString{String}, String}, !Matched::Integer) at strings/substring.jl:189
  repeat(::AbstractString, !Matched::Integer) at strings/basic.jl:659
Stacktrace:
 [1] docall(::Ptr{Nothing}) at /Users/aguang/CORE/scratch/JuliaCall/inst/julia/setup.jl:168
> julia_command('repeat(dna"TA",10)')
20nt DNA Sequence:
TATATATATATATATATATA
```

It seems that JuliaCall is treating the arguments as String inputs. In a similar vein using `julia_eval` on the arguments will work but I have to convert all the types it seems. (For example, 10 gets treated as ::Float64)

```{r}
> julia_call("DNASequence",julia_eval('dna"ACGT"'),julia_eval('dna"NNNN"'),julia_eval('dna"TGCA"'))
Julia Object of type BioSequence{DNAAlphabet{4}}.
ACGTNNNNTGCA
> julia_call("repeat", julia_eval('dna"TA"'),10)
 Show Traceback

 Rerun with Debug
 Error: Error happens in Julia.
MethodError: no method matching repeat(::BioSequence{DNAAlphabet{4}}, ::Float64)
Closest candidates are:
  repeat(::BioSequence{A<:Alphabet}, !Matched::Integer) where A<:Alphabet at /Users/aguang/.julia/packages/BioSequences/7i86L/src/bioseq/constructors.jl:96
Stacktrace:
 [1] docall(::Ptr{Nothing}) at /Users/aguang/CORE/scratch/JuliaCall/inst/julia/setup.jl:168

> julia_call("repeat", julia_eval('dna"TA"',need_return="Julia"),10)
 Show Traceback

 Rerun with Debug
 Error: Error happens in Julia.
MethodError: no method matching repeat(::BioSequence{DNAAlphabet{4}}, ::Float64)
Closest candidates are:
  repeat(::BioSequence{A<:Alphabet}, !Matched::Integer) where A<:Alphabet at /Users/aguang/.julia/packages/BioSequences/7i86L/src/bioseq/constructors.jl:96
Stacktrace:
 [1] docall(::Ptr{Nothing}) at /Users/aguang/CORE/scratch/JuliaCall/inst/julia/setup.jl:168
> julia_call("repeat", julia_eval('dna"TA"'),julia_eval('convert(Int,10)'))
Julia Object of type BioSequence{DNAAlphabet{4}}.
TATATATATATATATATATA
> julia_call("repeat", julia_eval('dna"TA"'),julia_eval('10'))
Julia Object of type BioSequence{DNAAlphabet{4}}.
TATATATATATATATATATA
```

Is there a way for `julia_call` to behave more like what would be expected in Julia with typing? One solution could be implementing a parameter `eval=TRUE` or `eval=FALSE` could be one solution to this. So for example if you did want to pass in ::String arguments you would do `eval=FALSE`, but if not you would provide `eval=TRUE`. Another solution could be having default behavior of `julia_call` to be evaluating the arguments. If the evaluation is undefined then it passes on the argument as a ::String instead.
