### 2019-03-06 #1284 @aguang
I've tested the code and read through the paper and most of it looks good to me. The package is very easy to use and even though I knew how much faster Julia was compared to R I was blown away by how much faster the Mandlebrot code was with JuliaCall. Very cool.

That said I've raised a few issues; 2 are to do with documentation and 1 is some code functionality.

Non-Contradiction/JuliaCall#91
Non-Contradiction/JuliaCall#92
Non-Contradiction/JuliaCall#93
