### 2019-03-06 Non-Contradiction/JuliaCall#91 @aguang
Linking to openjournals/joss-reviews#1284
