### 2019-03-06 Non-Contradiction/JuliaCall#92 @aguang
Linking to openjournals/joss-reviews#1284
