### 2019-03-06 Non-Contradiction/JuliaCall#93 @aguang
Linking to openjournals/joss-reviews#1284
