### 2019-03-06 Non-Contradiction/JuliaCall#91 @Non-Contradiction
I will add the reference to the `Troubleshooting` section in the `Basic Usage` section. Thanks for the suggestion.  BTW, `JuliaCall` should support the alias solution. This sounds like a bug to me. I will investigate more on the issue.
