### 2019-03-06 Non-Contradiction/JuliaCall#92 @Non-Contradiction
The difference is already documented. However, when I look at it again, the wording is not that clear. And other users have also asked me questions about these functions. So I think I need to improve the documentation and make it more clear. Thanks for the suggestion.
