### 2019-03-07 #1284 @leouieda
@aguang thank you for your comments :100:

@Non-Contradiction please reply to the issues that @matbesancon and @aguang have raised. If any changes need to be made, please let us know when you're done so we can have a second round of reviews. Please also look at the checklist items at the top of this issue to see which items are still missing.
