### 2019-03-07 687a24d
Add link of troubleshooting section in basic usage section at README for julia_setup problems, related to #91. [skip ci]
