+++ README.md
@@ -62,21 +62,24 @@ devtools::install_github("Non-Contradiction/JuliaCall")
 
 ## Basic Usage
 
+Before using `JuliaCall`, you need to do initial setup by function
+`julia_setup()` for automatic type conversion, Julia display systems,
+etc. It is necessary for every new R session to use the package. If not
+carried out manually, it will be invoked automatically before other
+`julia_xxx` functions. Solutions to some common error in `julia_setup()`
+are documented in the [troubleshooting
+section](#troubleshooting-and-way-to-get-help).
+
 ``` r
 library(JuliaCall)
-
-## Do initial setup for automatic type conversion, Julia display systems, etc. 
-## It is necessary for every new R session to use the package.
-## If not carried out manually, it will be invoked automatically before other julia_xxx functions.
-
 julia <- julia_setup()
 #> Julia version 1.0.3 at location /Applications/Julia-1.0.app/Contents/Resources/julia/bin will be used.
 #> Loading setup script for JuliaCall...
 #> Finish loading setup script for JuliaCall.
 
-## If you want to use Julia at a specific location, you could do the following:
-## julia_setup(JULIA_HOME = "the folder that contains Julia binary"), 
-## or you can set JULIA_HOME in command line environment or use `options(...)`
+## If you want to use `Julia` at a specific location, you could do the following:
+## julia_setup(JULIA_HOME = "the folder that contains Julia binary").
+## You can also set JULIA_HOME in command line environment or use `options(...)`.
 
 ## Different ways of using Julia to calculate sqrt(2)
 
