### 2019-03-07 e82d6dd
Improve documentations for julia_call, julia_eval and julia_command, related to #92.
