### 2019-03-07 Non-Contradiction/JuliaCall#93 @Non-Contradiction
This is in part due to the design of `julia_call`, whose purpose is to call functions in Julia on arguments in R (which will be converted to Julia object automatically).
When I created functions in `JuliaCall`, I wanted to make them somewhat corresponding to basic functions in R.For example, `julia_eval` corresponds to `eval`, `julia_help` corresponds to `help` and `julia_source` corresponds to `source`.
The R functionalities I had in mind for `julia_do.call` (yes, there is a function called this, but it may not be an appropriate name) and `julia_call` were `do.call` and `eval(call(name, ...))" in R. So `eval(call("round", 10.5))` corresponds to `julia_call("round", 10.5)", not `julia_call("round", "10.5")`.

The string evaluation is really a brilliant idea! Things like
```
julia_call("DNASequence",julia_eval('dna"ACGT"'),julia_eval('dna"NNNN"'),julia_eval('dna"TGCA"'))
```
definately doesn't look good.
However, I think it may not be a good idea to add an `eval` argument in `julia_call` since it makes the logic of `julia_call` more complicated and will lead to more potential confusion and bugs.
After more investigation, I think a string interpolation approach or something like that can be tried, for example, function [`str_interp`](shttps://stringr.tidyverse.org/reference/str_interp.html) in package `stringr` can interpolate R objects into strings, and we can follow a similar idea in `JuliaCall` to make some new function or utilize `julia_eval` and `julia_command` function to evaluate interpolated julia commands.

For example, if such function is implemented, then
```
a = 2;
julia_command("sqrt(${a})") ## which is equivalent to julia_command("sqrt(2)")
```
The logic here seems to be more clear and consistent than adding `eval` argument to `julia_call`.
And I need to explore more in related directions to have a good and performant implementation for this feature.

Currently, the most ideal way is to use things like
```
a = 2;
## again, this is quite similar to assign("aa", a) in R, except the assignment happens in Julia
julia_assign("aa", a)
julia_command("sqrt(aa)")
```

Does this answer the first part of the question?

Again, thank you very much for the suggestion!
