### 2019-03-07 Non-Contradiction/JuliaCall#93 @Non-Contradiction
And about the issue related to `10`, `typeof(10)` in R returns "double", so `10` actually represents a float number in R. In fact, `identical(10, 10.0)` returns `TRUE` in R, which means that `10` and `10.0` are programmatically undistinguishable in R itself. The implication is that the two R commands `abc(def, ..., 10)` and `abc(def, ..., 10.0)` are the same.

In sum, `JuliaCall` cannot distinguish between `julia_call("repeat", xxx,10)` from `julia_call("repeat", xxx,10.0)`. And I guess the most obvious and straightforward interpretation is `repeat(xxx, 10.0)` in Julia.
