### 2019-03-07 Non-Contradiction/JuliaCall#86 @Non-Contradiction
I merged #88 into master. After all, it only changes a little bit of functionality of the package, and I doubt any users will be negatively affected by this. So as far as I know, there is no deprecated usage of Julia now in the package.
