### 2019-03-07 Non-Contradiction/JuliaCall#92 @Non-Contradiction
Documentation about these functions has been improved, and it can be retrieved by `?julia_eval` and `?julia_command`.
