### 2019-03-08 Non-Contradiction/JuliaCall#93 @aguang
Ah, thanks for that, it does seem that `typeof` returns double for integers as well.

In terms of using a string interpolation approach, so would the examples I gave go from

```{r}
julia_call("convert", julia_eval("Float32"),1/3)
julia_call("DNASequence",julia_eval('dna"ACGT"'),julia_eval('dna"NNNN"'),julia_eval('dna"TGCA"'))
```

to

```{r}
julia_call("convert", ${"Float32"}, 1/3)
julia_call("DNASequence",${'dna"ACGT"'},${'dna"NNNN"'},${'dna"TGCA"'})
```

? Or would assignment need to happen first, i.e.

```{r}
float <- "Float32"
x <- 'dna"ACGT"'
y <- 'dna"NNNN"'
z <- 'dna"TGCA"'
julia_call("convert", ${float}, 1/3)
julia_call("DNASequence", ${x}, ${y}, ${z})
```

Either way, it seems like a good idea in terms of implementation to me. I would be happy to test it when you implement it.
