### 2019-03-08 Non-Contradiction/JuliaCall#91 @aguang
Looks good to me. No pressure on the alias issue, it is also trivial to make a symbolic link in /usr/local/bin or add Julia to my PATH.
