### 2019-03-08 #1284 @aguang
I closed the issues that were for functionality documentation and would recommend this paper for acceptance.

Quick question: There are 4 dois in the paper but whedon only checked 2. I checked the others and they are all OK but is this a bug? The other 2 are books that can be read in eBook form.
