### 2019-03-08 Non-Contradiction/JuliaCall#93 @Non-Contradiction
Oh, I forgot to mention that R's "real" integer looks like `10L`, but it is seldom used since most R functions just treat float numbers as integers.

The implementation currently in my mind would look like
```
julia_eval("convert(Float32}(${x})", x = 1/3)
julia_eval('DNASequence(dna"${x}", dna"${y}",dna"${z}"', x = "ACGT", y = "NNNN",  z = "TGCA")
```
or assignment can happen in environment first.

The idea is that when `julia_eval`, `julia_command` (or some new function for this functionality) find `${something}` in the command string, the functions will first evaluate `something` in R with the environment of the function, then interpolate the evaluation result into the command string, and finally evaluate the interpolated command string in Julia.
