### 2019-03-08 #1284 @leouieda
@aguang thank you conducting the review! :tada: I don't think there is a problem with the DOIs if you checked that they work manually. We can wait and see if Whedon will complain.

@matbesancon please let us know as soon as the remaining issues are resolved.
