### 2019-03-12 63ad577
Add position of benchmark code in paper. [skip ci]
Related to #94.
