### 2019-03-12 Non-Contradiction/JuliaCall#95 @matbesancon
[JOSS] Paper review
Last minor points on the paper itself:

```
There are R packages which wrap (no S)
for Disciplined Convex Programming ... providing (no by)
which utilizes DifferentialEquations.jl ->  using DifferentialEquations.jl
and the Julia package RCall
The package RJulia (Gong, Haverty...
```
