### 2019-03-12 #1284 @matbesancon
Closed most issues related to the software, one more regarding the paper:
https://github.com/Non-Contradiction/JuliaCall/issues/95
