### 2019-03-12 f09295a
Fix mistakes in paper, related to #95
