### 2019-03-12 Non-Contradiction/JuliaCall#95 @Non-Contradiction
Thank you very much! The problems are fixed by the commit f09295a.
