+++ appveyor.yml
@@ -2,15 +2,15 @@ environment:
   NOT_CRAN: "true"
   matrix:
   - JULIA_URL: "https://julialang-s3.julialang.org/bin/winnt/x64/0.6/julia-0.6-latest-win64.exe"
-    R_URL: "https://cran.rstudio.com/bin/windows/base/R-3.5.2-win.exe"
+    R_URL: "https://cran.rstudio.com/bin/windows/base/R-3.5.3-win.exe"
   - JULIA_URL: "https://julialang-s3.julialang.org/bin/winnt/x64/0.6/julia-0.6-latest-win64.exe"
     R_URL: "https://cran.rstudio.com/bin/windows/base/R-devel-win.exe"
   # - JULIA_URL: "https://julialang-s3.julialang.org/bin/winnt/x64/0.7/julia-0.7.0-win64.exe"
-  #   R_URL: "https://cran.rstudio.com/bin/windows/base/R-3.5.2-win.exe"
+  #   R_URL: "https://cran.rstudio.com/bin/windows/base/R-3.5.3-win.exe"
   - JULIA_URL: "https://julialang-s3.julialang.org/bin/winnt/x64/1.0/julia-1.0-latest-win64.exe"
-    R_URL: "https://cran.rstudio.com/bin/windows/base/R-3.5.2-win.exe"
+    R_URL: "https://cran.rstudio.com/bin/windows/base/R-3.5.3-win.exe"
   - JULIA_URL: "https://julialang-s3.julialang.org/bin/winnt/x64/1.1/julia-1.1-latest-win64.exe"
-    R_URL: "https://cran.rstudio.com/bin/windows/base/R-3.5.2-win.exe"
+    R_URL: "https://cran.rstudio.com/bin/windows/base/R-3.5.3-win.exe"
 
 notifications:
   - provider: Email
