### 2019-03-12 #1284 @matbesancon
All issues relevant are now closed, I recommend the paper for submission :)
