### 2019-03-13 #1284 @leouieda
Thank you @matbesancon and @aguang for all your work on the reviews! :1st_place_medal:

:tada: Congratulations @Non-Contradiction! :tada: Following the positive recommendation of the reviewers, I'm happy to accept your paper for publication in JOSS. We just need a couple of things from you now:

1. Double check the author list and affiliations
2. Make a new release of the software with latest changes after review and the updated JOSS paper text (this will be the official version published in JOSS)
3. Archive this new release on [Zenodo](https://zenodo.org/) or similar and post the DOI for the archive here

Please don't hesitate to ask if you need any help.
