### 2019-03-13 #1284 @Non-Contradiction
Thank you @matbesancon @aguang @leouieda for detailed reviews and constructive suggestions!

I checked the author list, make a new release on Github, and archive the release on Zenodo.
The DOI for the archive is DOI: 10.5281/zenodo.2591769
