### 2019-03-13 #1284 @leouieda
Thank you, @Non-Contradiction. I've run some last checks on the article proof and everything looks good to me.

@openjournals/joss-eics this submission is ready to be published.
