### 2019-03-13 #1284 @labarba
A few more editorial suggestions/requests:

- high-performance technique computing >> technique?
- Section 2.77 in (Xie, Allaire, & Grolemund, 2018) >>When the citation is part of the sentence, then use: Author (year). The citation should appear in parenthesis when it is not part of the sentence, e.g. “We take inspiration from Smith et al. (2006) to … The code implements Theory X (Jones, 2012) to obtain … “  For the syntax to obtain brackets in the right places, see: https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html
- There are R packages which wrap Julia packages >> remove syntactic expletive: "Some R packages wrap Julia packages…" Elsewhere in the paper, look for "there are" and see if you can remove it by rephrasing.
- The package XRJulia follows the extending R interface implemented by the XR package >> something missing? Check grammar.
- has a signi cant speed advantage >> suggest not to use "significant" unless you mean "statistically significant"
