### 2019-03-13 ed4e33e
Fixes in the paper.
