+++ paper/paper.md
@@ -18,11 +18,11 @@ bibliography: paper.bib
 
 `R` is a widely used software and computing environment for statistics,
 which provides a variety of statistical techniques and packages [@R].
-There are many `R` packages which are interfaces to other computer languages
+Many `R` packages provide interfaces to other computer languages
 to bring new functionalities to `R` or to accelerate computations,
 such as  `Rcpp` [@Rcpp] and `V8` [@V8].
 The package `JuliaCall` provides an `R` interface to `Julia`,
-which is a computer language for high-performance technique computing [@bezanson2017julia].
+which is a computer language for high-performance scientific computing [@bezanson2017julia].
 `JuliaCall` embeds `Julia` in `R`,
 and provides functions to evaluate `Julia` commands, to call `Julia` functions,
 to transmit data objects between `R` and `Julia`, and so on.
@@ -30,9 +30,9 @@ It also provides many utilities for user convenience.
 For example, `JuliaCall` gives detailed error messages for the embedded `Julia`.
 It also provides `Julia` package management functions such as installation and loading,
 and utility functions to get the documentation of `Julia` functions.
-`JuliaCall` can also be used in R Markdown document as the engine of `Julia` language, see Section 2.77 in [@xie2018r].
+`JuliaCall` can also be used in R Markdown document as the engine of `Julia` language, see Section 2.77 in @xie2018r.
 
-There are `R` packages which wrap `Julia` packages based on `JuliaCall` to provide new functionalities
+Some `R` packages wrap `Julia` packages based on `JuliaCall` to provide new functionalities
 or performance improvements to some existing packages in `R`.
 [`autodiffr`](https://github.com/Non-Contradiction/autodiffr) [@autodiffr] provides automatic differentiation to `R` functions by wrapping
 [`ForwardDiff.jl`](https://github.com/JuliaDiff/ForwardDiff.jl) and
@@ -41,18 +41,18 @@ or performance improvements to some existing packages in `R`.
 [`diffeqr`](https://github.com/JuliaDiffEq/diffeqr) [@diffeqr] solves differential equations in `R` using [`DifferentialEquations.jl`](http://docs.juliadiffeq.org/latest/).
 [`FixedEffectjlr`](https://github.com/eloualiche/FixedEffectjlr) [@fixedeffectjlr] estimates large fixed effects models in `R` by providing an interface to [`FixedEffectModels.jl`](https://github.com/matthieugomez/FixedEffectModels.jl).
 
-Besides `JuliaCall`, there are also some other interfaces between `R` and `Julia`:
+Besides `JuliaCall`, some other packages also provide interfaces between `R` and `Julia`:
 `R` packages [`XRJulia`](https://github.com/johnmchambers/XRJulia),
 [`RJulia`](https://github.com/armgong/rjulia),
 and the `Julia` package [`RCall.jl`](https://github.com/JuliaInterop/RCall.jl).
-The package `XRJulia` follows the extending `R` interface implemented by the `XR` package, which connects to `Julia` from `R` [@chambers2017extending].
+The package `XRJulia` connects to `Julia` from `R` [@chambers2017extending].
 It uses JavaScript Object Notation (JSON) format to transmit data, while `JuliaCall` copies objects in memory between `R` and `Julia`.
 It has performance disadvantages compared to `JuliaCall`.
 Table 1 depicts the times needed to transmit a $500\times 500$ matrix full of ones from `R` v3.5.2 to `Julia` v1.0.3 using `JuliaCall` v0.16.4 and `XRJulia` Github master b6224fa at the time of writing (there is no released version of `XRJulia` to support `Julia` v1.0 yet).
 The times are measured by `R` package `microbenchmark` with 1000 evaluation times.
[…]
