### 2019-03-13 #1284 @Non-Contradiction
Thank @labarba for the suggestions. They are of great help!
I have made the following changes accordingly.

* The wording of "high-performance technique computing" is from some introduction material of Julia by the Julia creators. I change it to high-performance scientific computing with similar meaning since it is easier to understand.
* I removed the brackets around the citation of xie2018r, from the link it seems to be the right syntax to give in-text citations. Previously I don't know these syntaxes. Thank you for the link.
* "there are" has been replaced.
* Extending R is an interface (or a framework of interfaces) connecting R to other languages, and the `XR` package implements the interface. `XRJulia` is an interface especially for Julia based on the `XR` package. Since this information of the `XRJulia` package can be easily found in the reference and not directly related to this paper, I removed the part and changed the sentence to "The package `XRJulia` connects to `Julia` from `R`."
* By "significant" I mean statistically significant. And I still removed the word since it is ambiguous.

Thank you for the suggestions again and hope these modifications make the paper more concise and clear.
