### 2019-03-13 #1284 @labarba
> Hi @Non-Contradiction — I'm going to ask you to edit the metadata of the Zenodo deposit (no need to get new version or new DOI) so the title and author list match the JOSS paper.

Looks like this is still pending?
