### 2019-03-13 #1284 @Non-Contradiction
I have questions in changing the metadata of the Zenodo deposit.
I used Zenodo's webhook to generate the deposit automatically from the release.
On the Zenodo website, it says creating a file with the name `.zenodo.json` can change the metadata.
But after creating the file, I found that the metadata didn't change automatically.
It seems that I have to create a new release to let Zenodo aware of the change?
