### 2019-03-13 #1284 @labarba
Congratulations, @Non-Contradiction, your JOSS paper is published!

Sincere thanks to the editor: @leouieda, and the reviewers: @matbesancon, @aguang 🙏
